requirements.txt
autopep8==1.7.0
buttercms-python==1.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
urllib3==1.25.10
Werkzeug==1.0.1
